334. Increasing Triplet Subsequence
we have three points first=nums[0], second, num=nums[i]

we try to make the first and second smaller, so that the third number will easy to be satisfied.

when we find num > second, then triplet found.
when we find first < num < second, then update second to make it smaller.
when we find num < first, then update first to make it smaller.