443. String Compression
We maintain two pointers, which are write and read pointers.

Write pointer is assigned to the write index of chars.

Read pointer is assigned to the read index of chars.

We assigned read to curr first, and find the next index which is not equals to chars[curr].

The Length will be read-curr.

Then write into the chars with write pointer.