394. Decode String
Stack solution:

when char is [ or letter, push the char into stack
when char is digits, find the whole digits and push into stack
when char is ], pop the char from stack until [. And use the times digits to multiply the char.
After construction of the string, push it into the stack.

Finally, the string in the stack will be the answer.