649. Dota2 Senate
Use two queues to line up radiant and dire party senates.

line up by their index. Offer both of the queues, small index will win the battle and re-offer to the queue with index+n to line in at the end of the queue to run next round.

Which queue is empty first will lose the battle.