437. Path Sum III
Depth first search + prefix sum:

depth first search each node and use a HashMap to store the sum -> occurrence of the pre nodes.

when find sum - targetSum prefix sum existing in HashMap, then find a path match the requirements.

recursively do the same algo on the left and right node of the node.