236. Lowest Common Ancestor of a Binary Tree
Depth first search solution:

if p or q is root return root
if p and q is not root, recursively find the lca of left and right subtree.
if lca of left subtree is null, return lca of right subtree.
if lca of right subtree is null, return lca of left subtree.
if lca of both left and right subtree are not null, curr node is the lca of the tree.